--- a/lib/Dialect/TritonGPU/Transforms/Prefetch.cpp
+++ b/lib/Dialect/TritonGPU/Transforms/Prefetch.cpp
@@ -159,6 +159,11 @@ LogicalResult Prefetcher::initialize() {
 
   for (triton::DotOp dot : dotsInFor) {
     auto kSize = dot.a().getType().cast<RankedTensorType>().getShape()[1];
+
+    // works better with nvidia tensor cores
+    unsigned elementWidth = dot.a().getType().cast<RankedTensorType>().getElementTypeBitWidth();
+    prefetchWidth = 256 / elementWidth;
+
     // Skip prefetching if kSize is less than prefetchWidth
     if (kSize < prefetchWidth)
       continue;