Use i32 as the storage type for <2xi16> and <4xi8>, as NVPTX inserts extra integer instructions for vector int types.

Performance before this PR: (8192x8192x8192-TN input)
bf16: 222 TFLOPS
i8:     339 TOPS

After this PR:
bf16: 272 TFLOPS
i8:      548 TOPS